# packages/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Music Repository API",
    "description": "The Music Repository API — the single source of truth for the HTTP contract.\nCompiles to OpenAPI 3.1 (`openapi.json`), which drives Orval (FE client + Zod) and the\nbackend Zod DTOs. Paths are declared ONLY here (never in NestJS decorators).",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/catalogue/items": {
      "get": {
        "operationId": "searchCatalogue",
        "description": "Browse & search the catalogue with faceted filters.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "explode": false
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "explode": false
          },
          {
            "name": "instrument",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "explode": false
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "explode": false
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "lesson",
                "song",
                "score",
                "exercise",
                "technique",
                "backing_track",
                "tool_page"
              ]
            },
            "explode": false
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogueList"
                }
              }
            }
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
    "/catalogue/items/{slug}": {
      "get": {
        "operationId": "getContentBySlug",
        "description": "Fetch a single catalogue item (with taxonomy + presigned media) by slug.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDetail"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
    "/catalogue/items/{slug}/related": {
      "get": {
        "operationId": "getRelatedContent",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelatedContent"
                }
              }
            }
          }
        },
        "tags": [
          "catalogue"
        ]
      }
    },
    "/classrooms/join": {
      "post": {
        "operationId": "joinClassroom",
        "description": "Join a classroom by its code.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classroom"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinClassroomInput"
              }
            }
          }
        }
      }
    },
    "/classrooms/{id}": {
      "get": {
        "operationId": "getClassroom",
        "description": "A classroom with its roster (owner or member only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDetail"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/archive": {
      "post": {
        "operationId": "archiveClassroom",
        "description": "Archive a classroom (owner only) — hides it from all rosters.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/assignments": {
      "get": {
        "operationId": "listAssignments",
        "description": "The class's assigned content (any owner or member).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentList"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ]
      },
      "post": {
        "operationId": "assignContent",
        "description": "Assign content to the class by slug (owner only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentList"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignContentInput"
              }
            }
          }
        }
      }
    },
    "/classrooms/{id}/assignments/{slug}": {
      "delete": {
        "operationId": "unassignContent",
        "description": "Remove an assignment (owner only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/grant-premium": {
      "post": {
        "operationId": "grantClassroomPremium",
        "description": "Grant premium to every current member of the classroom (owner only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDetail"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/leave": {
      "post": {
        "operationId": "leaveClassroom",
        "description": "Leave a classroom (a member removes themselves; the owner cannot leave — archive instead).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/members/{memberId}": {
      "delete": {
        "operationId": "removeClassroomMember",
        "description": "Remove a member from the classroom (owner only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/progress": {
      "get": {
        "operationId": "getClassProgress",
        "description": "Per-student progress across the class's assigned content (owner only).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassProgress"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ]
      }
    },
    "/classrooms/{id}/transfer": {
      "post": {
        "operationId": "transferClassroomOwnership",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          }
        },
        "tags": [
          "classrooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipInput"
              }
            }
          }
        }
      }
    },
...
    "/me/classrooms": {
      "get": {
        "operationId": "listClassrooms",
        "description": "Classrooms the current user owns or has joined.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomList"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ]
      },
      "post": {
        "operationId": "createClassroom",
        "description": "Create a classroom (the caller becomes its teacher/owner).",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Classroom"
                }
              }
            }
          }
        },
        "tags": [
          "classrooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignContentInput": {
        "type": "object",
        "required": [
          "contentSlug"
        ],
        "properties": {
          "contentSlug": {
            "type": "string"
          }
        }
      },
...
      "AssignmentList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Assignment"
            }
          }
        }
      },
...
      "CatalogueList": {
        "type": "object",
        "required": [
          "items",
          "facets",
          "total",
          "page",
          "pageSize"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentSummary"
            }
          },
          "facets": {
            "$ref": "#/components/schemas/Facets"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Paginated, faceted catalogue search result."
      },
...
      "ClassProgress": {
        "type": "object",
        "required": [
          "assignments",
          "members"
        ],
        "properties": {
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Assignment"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberProgress"
            }
          }
        }
      },
      "Classroom": {
        "type": "object",
        "required": [
          "id",
          "name",
          "joinCode",
          "memberCount",
          "role",
          "premiumGranted"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "joinCode": {
            "type": "string"
          },
          "memberCount": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ],
            "description": "The acting user's relationship to the classroom."
          },
          "premiumGranted": {
            "type": "boolean",
            "description": "True once the teacher has granted premium to the class."
          }
        }
      },
      "ClassroomDetail": {
        "type": "object",
        "required": [
          "id",
          "name",
          "joinCode",
          "memberCount",
          "role",
          "premiumGranted",
          "members"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "joinCode": {
            "type": "string"
          },
          "memberCount": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ],
            "description": "The acting user's relationship to the classroom."
          },
          "premiumGranted": {
            "type": "boolean",
            "description": "True once the teacher has granted premium to the class."
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassroomMember"
            }
          }
        }
      },
      "ClassroomList": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Classroom"
            }
          }
        }
      },
...
      "ContentDetail": {
        "type": "object",
        "required": [
          "slug",
          "title",
          "type",
          "visibility",
          "genres",
          "instruments",
          "topics",
          "tags",
          "media",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "lesson",
              "song",
              "score",
              "exercise",
              "technique",
              "backing_track",
              "tool_page"
            ]
          },
          "difficulty": {
            "type": "integer",
            "format": "int32"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "authed",
              "premium"
            ]
          },
          "tier": {
            "type": "string",
            "description": "For premium items: which plan unlocks it — \"premium\" or \"pro\"."
          },
          "locked": {
            "type": "boolean",
            "description": "True when this is premium content the current viewer is not entitled to (body/media withheld)."
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "instruments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "bodyMdx": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "license": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaAsset"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "description": "Full detail view of a catalogue item."
      },
      "ContentSummary": {
        "type": "object",
        "required": [
          "slug",
          "title",
          "type",
          "visibility",
          "genres",
          "instruments",
          "topics"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "lesson",
              "song",
              "score",
              "exercise",
              "technique",
              "backing_track",
              "tool_page"
            ]
          },
          "difficulty": {
            "type": "integer",
            "format": "int32"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "authed",
              "premium"
            ]
          },
          "tier": {
            "type": "string",
            "description": "For premium items: which plan unlocks it — \"premium\" or \"pro\"."
          },
          "locked": {
            "type": "boolean",
            "description": "True when this is premium content the current viewer is not entitled to (body/media withheld)."
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "instruments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          },
          "topics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyRef"
            }
          }
        },
        "description": "List/card view of a catalogue item."
      },
...
      "CreateClassroomInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "JoinClassroomInput": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "URI identifying the problem category."
          },
          "title": {
            "type": "string",
            "description": "Short, stable, human summary of the problem type."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "description": "Human explanation of this occurrence."
          },
          "instance": {
            "type": "string",
            "description": "URI identifying this specific occurrence."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code — the programmatic contract."
          },
          "traceId": {
            "type": "string",
            "description": "W3C trace id for support correlation."
          },
          "docUrl": {
            "type": "string",
            "description": "Link to human documentation for this code."
          }
        },
        "description": "RFC 9457 Problem Details (`application/problem+json`)."
      },
...
      "RelatedContent": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentSummary"
            }
          }
        },
        "description": "Related items — published content sharing genre/instrument/topic, ranked by overlap."
      },
...
      "TaxonomyRef": {
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "description": "A taxonomy reference (genre / instrument / topic / tag)."
      },
...
      "TransferOwnershipInput": {
        "type": "object",
        "required": [
          "memberId"
        ],
        "properties": {
          "memberId": {
            "type": "string"
          }
        },
        "description": "Transfer ownership to a current member (owner only)."
      }
    }
  }
}